rename package to photonstream
and auto-remove trailing spaces
