+++ src/main/java/fact/photonstream/ConvertSinglePulses2Timeseries.java
@@ -1,4 +1,4 @@
-package fact.utils;
+package fact.photonstream;
 
 import fact.features.singlePulse.timeSeriesExtraction.AddFirstArrayToSecondArray;
 import fact.features.singlePulse.timeSeriesExtraction.SinglePulseExtractor;
@@ -48,8 +48,8 @@ public class ConvertSinglePulses2Timeseries implements Processor {
             // Add the single pulses to the time series
             for (int pulse = 0; pulse < singlePulses[pix].length; pulse++) {
                 AddFirstArrayToSecondArray.at(
-                    pulseTemplate, 
-                    currentTimeSeries, 
+                    pulseTemplate,
+                    currentTimeSeries,
                     singlePulses[pix][pulse]);
             }
 
@@ -61,7 +61,7 @@ public class ConvertSinglePulses2Timeseries implements Processor {
 
         SinglePulseExtractor.Config config = new SinglePulseExtractor.Config();
         timeSeries = ElementWise.multiply(
-            timeSeries, 
+            timeSeries,
             config.factSinglePeAmplitudeInMv);
 
         input.put(timeSeriesKey, timeSeries);
