+++ src/main/java/fact/photonstream/PhotonStream2ArrivalTime.java
@@ -1,4 +1,4 @@
-package fact.utils;
+package fact.photonstream;
 
 import java.util.Arrays;
 import stream.Data;
