+++ src/main/java/fact/photonstream/PhotonStream2NumberOfPhotons.java
@@ -1,4 +1,4 @@
-package fact.utils;
+package fact.photonstream;
 
 import stream.Data;
 import stream.Processor;
