+++ src/main/java/fact/photonstream/singlePulse/ArrivalTimeFromSlope.java
@@ -1,7 +1,7 @@
 /**
- * 
+ *
  */
-package fact.features.singlePulse;
+package fact.photonstream.singlePulse;
 
 
 import fact.Utils;
@@ -16,9 +16,9 @@ import java.util.ArrayList;
 /**
  * Finds pulse arrival time by finding the maximum slope in the leading edges of pulses or by finding the beginning of the pulse.
  * Also can find values to use for a baseline for calculating individual pulse sizes
- * 
+ *
  *@author Katie Gray &lt;kathryn.gray@tu-dortmund.de&gt;
- * 
+ *
  */
 public class ArrivalTimeFromSlope implements Processor {
     static Logger log = LoggerFactory.getLogger(ArrivalTimeFromSlope.class);
@@ -33,17 +33,17 @@ public class ArrivalTimeFromSlope implements Processor {
         //positions of arrival times
     @Parameter(required = false)
     private String visualizeKey;
-        //array of size data.length with values of zero except for time slices of arrival times. 
+        //array of size data.length with values of zero except for time slices of arrival times.
     @Parameter(required = false)
     private String baselineKey;
         //used in OpenShutterPulseSize to account for negative values
- 
+
     private int skipFirstSlices = 0;
         //start searching after this number of slices
-    private int skipLastSlices = 0;   
-        //stop searching this many slices before the end of the timeline 
+    private int skipLastSlices = 0;
+        //stop searching this many slices before the end of the timeline
     private int width = 1;
-        //should be an odd number. 
+        //should be an odd number.
 
     private int npix;
 
@@ -55,13 +55,13 @@ public class ArrivalTimeFromSlope implements Processor {
             width++;
             log.info("ArrivalTimeFromSlope only supports odd window lengths. New length is: " + width);
         }
-        
+
         double[] data = (double[]) input.get(key);
         double[] slopes = (double[]) input.get(derivationKey);
         int roi = data.length / npix;
-        
+
         ArrayList[] pulsePeaks =  new ArrayList[npix];
-            //the position where pulse leading edges end 
+            //the position where pulse leading edges end
         int[][] arrivalTimes = new int[npix][];
             //arrival times for all pulses in each pixel
         double[][] baselineValues = new double[npix][];
@@ -72,21 +72,21 @@ public class ArrivalTimeFromSlope implements Processor {
         for(int i = 0; i < data.length; i++){
             visualizePositions[i] = 0;
         }
-                
+
         //for each pixel
-        for (int pix = 0; pix < npix; pix++) {          
+        for (int pix = 0; pix < npix; pix++) {
             pulsePeaks[pix] = findPulsePeaks(pix, roi, slopes);
 
 		arrivalTimes[pix] 	= new int[pulsePeaks[pix].size()];
 		baselineValues[pix] 	= new double[pulsePeaks[pix].size()];
             arrivalTimes[pix] = findArrivalTimes(pix, roi, width, data, slopes, pulsePeaks, visualizePositions, baselineValues);
         }
-        
+
         input.put(outputKey, arrivalTimes);
         input.put(visualizeKey, visualizePositions);
         input.put(baselineKey, baselineValues);
 //        System.out.println(Arrays.toString(baselineValues));
-        
+
 
         return input;
     }
@@ -95,27 +95,27 @@ public class ArrivalTimeFromSlope implements Processor {
      * @param pix - Pixel to check
      * @param roi - the number of slices in one event
      */
-    
+
  //the function that finds the pulses. returns positions of the peaks
     public ArrayList findPulsePeaks(int pix, int roi, double[] slopes){
-    
+
         ArrayList<Integer> peaks = new ArrayList<Integer>();
         int risingEdgeLength = 10;
 
-        
+
         for(int slice=0+skipFirstSlices; slice < roi-skipLastSlices; slice++){
             int pos = pix*roi+slice;
             boolean peak = true;
             boolean check = false;
                 //allows one slice to have a negative slope in the leading edge
-            
 
-            
+
+
             if(slopes[pos] <= 0){
                 peak = false;
                 continue;
             }
-            
+
             else{
                 for(int i=0; i < risingEdgeLength; i++){
                     if(slice-i >= 0 && slice + i < roi){
@@ -136,11 +136,11 @@ public class ArrivalTimeFromSlope implements Processor {
                         break;
                     }
                 }
-            }                                           
-                
+            }
+
             if(peak == false) {continue;}
-                
-                int k=0;                
+
+                int k=0;
                 while(slice+k < roi){
                     if(slopes[pos+k] > 0){
                         k++;
@@ -149,43 +149,43 @@ public class ArrivalTimeFromSlope implements Processor {
                     else break;
                 }
                 slice += k - 1;
-                    
+
             if(peak == true){
                 peaks.add(slice);
             }
         }
-        
+
         return peaks;
     }
-          
-    
+
+
 //the function that finds the starting point of the pulse, defined by the first position with a positive slope, and
-//the position of maximum slope. both values can be used for arrival time or baseline values   
+//the position of maximum slope. both values can be used for arrival time or baseline values
 	public int[] findArrivalTimes(int pix, int roi, int width, double[] data, double[] slopes, ArrayList[] pulsePeaks, double[] visualizePositions, double[][] baselineValues){
 		ArrayList<Integer> times = new ArrayList<Integer>();
 		ArrayList<Double> baseValues = new ArrayList<Double>();
 		ArrayList<Integer> peaks = pulsePeaks[pix];
 		int number = peaks.size();
 		int pivot = (int) (width/2.0);
-		
+
 		for(int pulse = 0; pulse < number; pulse++){
 			int end = (Integer) peaks.get(pulse);
-		
+
 			//find the starting point of the leading edge
 			int current = end;
 			while(slopes[pix*roi+current-1] > 0){
 				current --;
-			}				
+			}
 			int start = current;		//start is the first position of the leading edge
-			
+
 //			accounting for 'false positives':
 			if(pix*roi+end < data.length && pix*roi+start > 0){
-				double difference = data[pix*roi+end] - data[pix*roi+start];			
+				double difference = data[pix*roi+end] - data[pix*roi+start];
 				if(difference < 7){
 					continue;
 				}
 			}
-	
+
 	//find max slope over leading edge
 			int maxpos = 0;
 			double maxslope = 0;
@@ -195,11 +195,11 @@ public class ArrivalTimeFromSlope implements Processor {
 				if(width == 1){
 					double currentslope = slopes[pos];
 					if(currentslope > maxslope){
-						maxslope = currentslope; 
+						maxslope = currentslope;
 						maxpos = slice;
 					}
 				}
-				
+
 				else{
 					if(slice+pivot < end && slice-pivot > start){
 						double currentslope = data[pos+pivot] - data[pos-pivot];
@@ -209,21 +209,21 @@ public class ArrivalTimeFromSlope implements Processor {
 						}
 					}
 				}
-			} 
-			
+			}
+
 			if(start > 0+skipFirstSlices && end < roi-skipLastSlices && end - maxpos < 14){
 				visualizePositions[pix*roi+start] = 15;
 				times.add(start);
 				baseValues.add(data[pix*roi+start]);
 			}
-				//to use maximum slope instead of first position of rising edge, simply replace start with maxpos. 
+				//to use maximum slope instead of first position of rising edge, simply replace start with maxpos.
 		}
 		baselineValues[pix] = Utils.arrayListToDouble(baseValues);
         return Utils.arrayListToInt(times);
     }
-    
 
-     
+
+
     /*
      * Getters and Setters
      */
@@ -268,7 +268,7 @@ public class ArrivalTimeFromSlope implements Processor {
     public void setWidth(int width) {
         this.width = width;
     }
-    
+
     public String getBaselineKey() {
         return baselineKey;
     }
@@ -293,5 +293,5 @@ public class ArrivalTimeFromSlope implements Processor {
         this.skipLastSlices = skipLastSlices;
     }
 
-    
+
 }
