+++ src/main/java/fact/photonstream/singlePulse/FWHMPulses.java
@@ -1,4 +1,4 @@
-package fact.features.singlePulse;
+package fact.fephotonstreamatures.singlePulse;
 
 import fact.Constants;
 import fact.Utils;
