+++ src/main/java/fact/photonstream/singlePulse/FindThresholdCrossings.java
@@ -1,4 +1,4 @@
-package fact.features.singlePulse;
+package fact.photonstream.singlePulse;
 
 import fact.Constants;
 import fact.Utils;
@@ -11,12 +11,12 @@ import stream.annotations.Parameter;
 import java.util.ArrayList;
 
 /**
- * 
- * Calculates the positions of pulse candidates by finding threshold crossings with boundary restrictions. 
+ *
+ * Calculates the positions of pulse candidates by finding threshold crossings with boundary restrictions.
  * Outputs an array containing a list of positions for each pixel.
- * 
+ *
  * @author Katie Gray &lt;kathryn.gray@tu-dortmund.de&gt;
- * 
+ *
  */
 
 public class FindThresholdCrossings implements Processor {
@@ -28,12 +28,12 @@ public class FindThresholdCrossings implements Processor {
     private String outputKey;
     @Parameter(required = false)
     private String visualizeOutputKey;
-    
+
     private int minBelow = 0;
     private int minAbove = 0;
-    
-    private double threshold = 5; 
-    
+
+    private double threshold = 5;
+
     /*
      * minBelow is minimum number of slices required to be below the threshold before the threshold crossing
      * minAbove is minimum number of slices required to be above the threshold after the threshold crossing
@@ -41,8 +41,8 @@ public class FindThresholdCrossings implements Processor {
      *
      */
     private int npix;
-    
-    
+
+
     @Override
     public Data process(Data input) {
         Utils.isKeyValid(input, "NPIX", Integer.class);
@@ -50,17 +50,17 @@ public class FindThresholdCrossings implements Processor {
 
         double[] data = (double[]) input.get(key);
         int roi = data.length / npix;
-        
+
         //the graph of the array positions visually shows the positions of the crossings for an individual pixel
         double[] positions =  new double[data.length];
-        
+
         //the array CrossPositions contains lists of positions of crossings for each pixel
         int[][] CrossPositions = new int[npix][];
 
-            for(int pix=0; pix < npix; pix++){              
+            for(int pix=0; pix < npix; pix++){
                 //creates a list of positions of threshold crossings for a single pixel
                 ArrayList<Integer> slices = new ArrayList<Integer>();
-                
+
                 for(int slice=0; slice < roi; slice++){
                     if(candidate(pix, slice, roi, data) == true) {
                         slices.add(slice);
@@ -71,26 +71,26 @@ public class FindThresholdCrossings implements Processor {
                         positions[pix*roi+slice]=0;
                     }
                 }
-                
-                //  if no crossing, list for that pixel is empty                
+
+                //  if no crossing, list for that pixel is empty
                 CrossPositions[pix] = new int[slices.size()];
                 CrossPositions[pix] = Utils.arrayListToInt(slices);
             }
-         
+
             input.put(visualizeOutputKey, positions);
             input.put(outputKey, CrossPositions);
             return input;
-        
+
     }
-    
+
     //to determine threshold crossings
     public boolean candidate(int pix, int slice, int roi, double[] data){
         boolean answer = true;
 
         int pos = pix * roi + slice;
-        
+
         if(data[pos] < threshold){return false;}
-    
+
         for(int i=1; i < minBelow; i++){
             if(slice-i < 0){return false;}
             if(data[pos-i] >= threshold){return false;}
@@ -98,15 +98,15 @@ public class FindThresholdCrossings implements Processor {
 
         for(int k=0; k < minAbove; k++){
             if(slice+k > roi){return false;}
-            if(pos+k == npix * roi) {return false;}    
+            if(pos+k == npix * roi) {return false;}
             if(data[pos+k] <= threshold){return false;}
             }
-    
+
         return answer;
     }
 
     //Getters and Setters
-    
+
     public String getKey() {
         return key;
     }
@@ -154,7 +154,7 @@ public class FindThresholdCrossings implements Processor {
     public void setThreshold(double threshold) {
         this.threshold = threshold;
     }
-    
-    
+
+
 
 }
