+++ src/main/java/fact/photonstream/singlePulse/OpenShutterPulseSize.java
@@ -1,7 +1,7 @@
 /**
- * 
+ *
  */
-package fact.features.singlePulse;
+package fact.photonstream.singlePulse;
 
 import fact.Utils;
 import org.slf4j.Logger;
@@ -14,9 +14,9 @@ import java.util.ArrayList;
 
 /**
  * Finds the integral of pulses by defining the specific baseline for each pulse to account for negative signal
- * 
+ *
  *@author Katie Gray &lt;kathryn.gray@tu-dortmund.de&gt;
- * 
+ *
  */
 public class OpenShutterPulseSize implements Processor {
     static Logger log = LoggerFactory.getLogger(OpenShutterPulseSize.class);
@@ -37,7 +37,7 @@ public class OpenShutterPulseSize implements Processor {
         //number of slices over which we integrate
 
     private int npix;
-    
+
     @Override
     public Data process(Data input) {
         Utils.isKeyValid(input, "NPIX", Integer.class);
@@ -47,15 +47,15 @@ public class OpenShutterPulseSize implements Processor {
 		int[][] arrivalTimes = (int[][]) input.get(arrivalTimeKey);
 		double[][] baselineValues = (double[][]) input.get(baselineKey);
 	    double[][] pulseSizes = new double[npix][];
-      
+
 		//for each pixel
 		for (int pix = 0; pix < npix; pix++) {
 			pulseSizes[pix] = new double[arrivalTimes[pix].length];
 			pulseSizes[pix] = calculateSizes(pix, roi, data, arrivalTimes, baselineValues);
 		}
-		
+
         input.put(outputKey, pulseSizes);
-        
+
         return input;
     }
 
@@ -65,26 +65,26 @@ public class OpenShutterPulseSize implements Processor {
      * @param data the array which to check
      * @return
      */
-	
+
     public double[] calculateSizes(int pix, int roi, double[] data, int[][] arrivalTimes, double[][] baselineValues){
       //changed from int to double
 		ArrayList<Double> sizes = new ArrayList<Double>();
-    	
+
         if(arrivalTimes[pix].length > 0){
         	int numberPulses = arrivalTimes[pix].length;
-        	
+
         	//return an empty list for any pixel where the number of pulses is not equal to the number of baseline values
         	if(numberPulses != baselineValues[pix].length){
         		System.out.println("Error - arrival times don't match up with baseline values");
         		return Utils.arrayListToDouble(sizes);
         	}
-        	
+
         	for(int i = 0; i < numberPulses; i++){
                   double integral = 0;
                   int start =  arrivalTimes[pix][i];
                   double baseline =  baselineValues[pix][i];
-                  
-                
+
+
                   //ignore pulses that are too close together
                   if(numberPulses > 1 && i != 0){
                 	 int first =  arrivalTimes[pix][i-1];
@@ -94,17 +94,17 @@ public class OpenShutterPulseSize implements Processor {
                           continue;
                       }
                   }
-                  
+
                   for(int slice = start; slice < start + width; slice++){
                        int pos = pix * roi + slice;
                        integral += (data[pos] - baseline);
                   }
                   if(integral > 0) sizes.add(integral);
-            }       
+            }
         }
         return Utils.arrayListToDouble(sizes);
-    }   
-     
+    }
+
     /*
      * Getters and Setters
      */
@@ -148,6 +148,6 @@ public class OpenShutterPulseSize implements Processor {
     public void setWidth(int width) {
         this.width = width;
     }
-    
+
 
 }
