+++ src/main/java/fact/photonstream/singlePulse/PulseMaxAmplitude.java
@@ -1,7 +1,7 @@
 /**
- * 
+ *
  */
-package fact.features.singlePulse;
+package fact.photonstream.singlePulse;
 
 import fact.Constants;
 import fact.Utils;
@@ -14,11 +14,11 @@ import stream.annotations.Parameter;
 import java.util.ArrayList;
 
 /**
- * This processor calculates the position of the maximum value for each pulse in each pixel. 
+ * This processor calculates the position of the maximum value for each pulse in each pixel.
  * Input and output are both arrays of size NUMBEROFPIXEL with lists of positions for each pixel.
- * 
- *modified by Katie Gray (kathryn.gray@tu-dortmund.de) from MaxAmplitudePosition 
- * 
+ *
+ *modified by Katie Gray (kathryn.gray@tu-dortmund.de) from MaxAmplitudePosition
+ *
  */
 public class PulseMaxAmplitude implements Processor {
     static Logger log = LoggerFactory.getLogger(PulseMaxAmplitude.class);
@@ -42,7 +42,7 @@ public class PulseMaxAmplitude implements Processor {
 		int[][] pulsePositions = (int[][]) input.get(pulsePositionKey);
         int roi = data.length / npix;
         int[][] positions =  new int[npix][];
-        
+
 		//for each pixel
 		for (int pix = 0; pix < npix; pix++) {
 			positions[pix] = new int[pulsePositions[pix].length];
@@ -55,15 +55,15 @@ public class PulseMaxAmplitude implements Processor {
     }
 
     /**
-     * finds the position of the highest value in the pulse. if max is not unique, last position will be taken. 
+     * finds the position of the highest value in the pulse. if max is not unique, last position will be taken.
      * @param pix Pixel to check
      * @param roi Basically the number of slices in one event
      * @param data the array which to check
      * @return
      */
-	
+
     public int[] findMaximumPositions(int pix, int roi, double[] data, int[][] pulsePositions){
-      
+
         ArrayList<Integer> maxima = new ArrayList<Integer>();
 
         if(pulsePositions[pix].length > 0){
@@ -77,8 +77,8 @@ public class PulseMaxAmplitude implements Processor {
                        if(slice > roi) {break;}
                        if(pos == data.length) {break;}
                        double value = data[pos];
-                        //update maxvalue and position if current value exceeds old value 
-                       if(slice != start && slice != start + 30){ 
+                        //update maxvalue and position if current value exceeds old value
+                       if(slice != start && slice != start + 30){
                            if(value >= tempMaxValue){
                                tempMaxValue = value;
                                Position = slice;
@@ -88,11 +88,11 @@ public class PulseMaxAmplitude implements Processor {
                   maxima.add(Position);
             }
         }
-        
+
         return Utils.arrayListToInt(maxima);
     }
-          
-     
+
+
     /*
      * Getters and Setters
      */
