+++ src/main/java/fact/photonstream/singlePulse/timeSeriesExtraction/ElementWise.java
@@ -1,4 +1,4 @@
-package fact.features.singlePulse.timeSeriesExtraction;
+package fact.photonstream.singlePulse.timeSeriesExtraction;
 
 /**
  * A collection of element-wise operations on double[] arrays with a scalar.
